[instagram_type] move
