+++ instagram_to_discord/discord_event_listener.py
@@ -7,9 +7,9 @@ from debug import DEBUG
 from instagram_to_discord.sites.instagram.instagram import send_instagram_images_from_cache_for_specified_index
 
 from .cookie_requests import requests_get_cookie
-from .instagram_type import get_multiple_medias_from_str
 from .sites.tiktok_handler import handle_tiktok_main
 from .sites.youtube_handler import handle_youtube_main
+from .sites.instagram.instagram_type import get_multiple_medias_from_str
 from .sites.instagram.instagram_process import process_instagram
 from .sites.twitter.twitter import send_twitter_images_from_cache_for_specified_index, twitter_line_to_image_urls
 from .sites.twitter.twitter_process import process_twitter
