+++ instagram_to_discord/sites/instagram/instagram_type.py
